--- a/.github/workflows/cypress.yml
+++ b/.github/workflows/cypress.yml
@@ -25,6 +25,10 @@ jobs:
       - name: Set up Node environment
         uses: ./.github/actions/setup-node-env
 
+      # https://github.com/cypress-io/github-action/issues/1044
+      - run: pnpm cypress install
+        working-directory: dotcom-rendering
+
       - name: Cypress run
         uses: cypress-io/github-action@v6
         with: