--- a/.github/workflows/prettier.yml
+++ b/.github/workflows/prettier.yml
@@ -11,4 +11,4 @@ jobs:
         uses: ./.github/actions/setup-node-env
 
       - name: Prettier check
-        run: yarn prettier:check
+        run: pnpm prettier:check