--- a/.husky/pre-commit
+++ b/.husky/pre-commit
@@ -6,4 +6,4 @@
 
 . "$(dirname "$0")/_/husky.sh"
 
-yarn lint-staged
+pnpm lint-staged