--- a/.npmrc
+++ b/.npmrc
@@ -1,0 +1,1 @@
+save-prefix=''