--- a/.prettierignore
+++ b/.prettierignore
@@ -1,7 +1,7 @@
 dist
-.yarn
 deno
 .nvmrc
+pnpm-lock.yaml
 
 # generated files
 dotcom-rendering/src/model/*-schema.json