--- a/.yarnrc.yml
+++ b/.yarnrc.yml
@@ -1,5 +1,0 @@
-defaultSemverRangePrefix: ''
-
-nmHoistingLimits: workspaces
-
-nodeLinker: node-modules