--- a/README.md
+++ b/README.md
@@ -22,11 +22,11 @@ corepack enable
 
 ## Install
 
-Run `yarn` in the root directory of this project to install packages.
+Run `pnpm` in the root directory of this project to install packages.
 
 ## Run
 
-You should always `cd` into the correct subdirectory before running commands (e.g `make dev` for dotcom-rendering, or `yarn watch` for apps-rendering) except for storybook.
+You should always `cd` into the correct subdirectory before running commands (e.g `make dev` for dotcom-rendering, or `pnpm watch` for apps-rendering) except for storybook.
 
 ### `apps rendering`
 
@@ -42,10 +42,10 @@ Most commands are run from within each project but the following are managed fro
 
 ### Storybook/Chromatic
 
-`yarn storybook` - Runs Storybook for all projects
-`yarn build-storybook` - Builds Storybook for all projects
+`pnpm storybook` - Runs Storybook for all projects
+`pnpm build-storybook` - Builds Storybook for all projects
 
-Chromatic now runs at project level. `cd` into the project dir and run `yarn chromatic -t [CHROMATIC PROJECT TOKEN]`
+Chromatic now runs at project level. `cd` into the project dir and run `pnpm chromatic -t [CHROMATIC PROJECT TOKEN]`
 
 You can find the token in the project Chromatic instance.