--- a/apps-rendering/README.md
+++ b/apps-rendering/README.md
@@ -8,7 +8,7 @@
 4. Install dependencies:
 
 ```sh
-yarn
+pnpm
 ```
 
 ### Run (Development)
@@ -20,7 +20,7 @@ yarn
 This is the simplest way to get started, but will intermingle all the logs together in one shell. If you're doing development work it might be easier to run the client and server in separate shells as described in the next subsection.
 
 ```sh
-yarn watch
+pnpm watch
 ```
 
 View in a browser at http://localhost:3030.
@@ -58,13 +58,13 @@ This will output each command's logs to different shells, which can make develop
 To start the server:
 
 ```sh
-yarn watch:server
+pnpm watch:server
 ```
 
 To start the client:
 
 ```sh
-yarn watch:client
+pnpm watch:client
 ```
 
 View in a browser at http://localhost:3030
@@ -76,13 +76,13 @@ _**Note**: You will need to refresh the page to see any changes you make to the 
 You may need to build the client-side code first with:
 
 ```sh
-yarn build:client
+pnpm build:client
 ```
 
 Then start the server with:
 
 ```sh
-yarn watch:server
+pnpm watch:server
 ```
 
 View in a browser at http://localhost:3040
@@ -94,7 +94,7 @@ _**Note**: You will need to refresh the page to see any changes you make to the 
 The unit tests are built using Jest. They can be run with the following command:
 
 ```sh
-yarn test
+pnpm test
 ```
 
 ### Lint
@@ -102,13 +102,13 @@ yarn test
 ESLint is used to validate the code. It can be run like this:
 
 ```sh
-yarn lint
+pnpm lint
 ```
 
 It can automatically fix problems for you:
 
 ```sh
-yarn lint:fix
+pnpm lint:fix
 ```
 
 ### Storybook
@@ -116,7 +116,7 @@ yarn lint:fix
 A good way to see components and test them in isolation is to run storybook:
 
 ```sh
-yarn storybook
+pnpm storybook
 ```
 
 Stories are deployed on [GitHub pages](https://guardian.github.io/apps-rendering)