--- a/dotcom-rendering/cdk.json
+++ b/dotcom-rendering/cdk.json
@@ -1,5 +1,5 @@
 {
-	"app": "yarn tsx cdk/bin/cdk.ts",
+	"app": "pnpm tsx cdk/bin/cdk.ts",
 	"context": {
 		"aws-cdk:enableDiffNoFail": "true",
 		"@aws-cdk/core:stackRelativeExports": "true"