--- a/dotcom-rendering/docs/contributing/where-should-my-code-live.md
+++ b/dotcom-rendering/docs/contributing/where-should-my-code-live.md
@@ -2,7 +2,7 @@
 
 <!-- START doctoc generated TOC please keep comment here to allow auto update -->
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
-<!-- Automatically created by running `yarn createtoc` in a pre-commit hook -->
+<!-- Automatically created by running `pnpm createtoc` in a pre-commit hook -->
 
 -   [File locations](#file-locations)
 -   [File naming schemes](#file-naming-schemes)