--- a/dotcom-rendering/tsconfig.json
+++ b/dotcom-rendering/tsconfig.json
@@ -12,7 +12,7 @@
 		"noUncheckedIndexedAccess": true,
 		"paths": {
 			/* Aliases should also be added to the webpack, babel and jest configurations */
-			"*": ["node_modules/@types/*", "*"] // Make sure that yarn linking doesn't confuse things https://github.com/microsoft/TypeScript/issues/11916#issuecomment-257130001
+			"*": ["node_modules/@types/*", "*"] // Make sure that package linking doesn't confuse things https://github.com/microsoft/TypeScript/issues/11916#issuecomment-257130001
 		},
 		"preserveConstEnums": true
 	},