--- a/dotcom-rendering/webpack/webpack.config.server.js
+++ b/dotcom-rendering/webpack/webpack.config.server.js
@@ -45,7 +45,7 @@ module.exports = {
 					nodeExternals({
 						allowlist: [/^@guardian/],
 						additionalModuleDirs: [
-							// Since we use yarn-workspaces for the monorepo, node_modules will be co-located
+							// Since we use workspaces for the monorepo, node_modules will be co-located
 							// both in the '(project-root)/dotcom-rendering/node_modules' directory (default for webpack-node-externals)
 							// but also in project root, and any workspaces we link to.
 							// We want to make sure all of these are removed from the server build.