… apps-rendering/package.json
… dotcom-rendering/package.json
… pnpm-lock.yaml
… scripts/preinstall.mjs
… yarn.lock