## What does this change?

### Swaps `yarn` for `pnpm`

`pnpm` is much faster to install, much lighter on disk, and because of that it is able isolate dependencies so that no transitive deps are available. 

Updating the apps to get this working has surfaced real bugs and instability. 

Using `pnpm` should hopefully prevent this class of bug recurring, while keeping a useable, predictable `node_modules` dirs around (unlike, for example, [yarn's pnp mode](https://yarnpkg.com/features/pnp), which solves the problem in a more non-standard way).

> [!NOTE]
> To get the build working, I had to create the `pnpm-lock.yaml` from scratch – importing the current `yarn.lock` resulted in various issues with webpack types (and i think others too) which starting clean resolved.
>
> This means that dependencies will have been bumped to the latest available versions in the specified ranges. This is probably a good thing, but it does mean that build output is not going to be identical to `main`. I _think_ that's ok?

| Installing `node_modules` | Before | After |
| ----------- | ----------- | ---------- |
| `make install` (no changes) | 2.9s | 1.7s |
| `./.github/actions/setup-node-env` (with a cache hit) | 51s | 20s |

### Removes the `preinstall` corepack check

Only yarn runs the `preinstall` hook _before_ installing (see [here](https://github.com/pnpm/pnpm/issues/3836) and [here](https://github.com/npm/cli/issues/2660)), so there's not a lot of point in keeping this around.