@sndrs 2024-01-02
> Just as a precaution could we deploy this to CODE pls to make sure the prod build and start is not affected. I will raise a ticket to add a GHA check to run the prod task so this is automated in future.

DCR deployed to CODE here https://riffraff.gutools.co.uk/deployment/view/0d3751e5-0a93-4d50-8961-5fb5c7cd0ec3

checking AR...